+++ lib/pdf/models/pdf_invoice_model.dart
@@ -20,6 +20,7 @@ class PdfInvoiceModel extends PdfItemizedInvoiceModel {
     super.discount,
     super.tax,
     super.paymentMethod,
+    super.dueDate,
     super.total,
     super.paidAmount,
   }) : super(